requirements.txt
aniso8601==7.0.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-RESTful==0.3.7
get==2019.4.13
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
JsonForm==0.0.2
jsonify==0.5
jsonschema==3.0.1
JsonSir==0.0.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
post==2019.4.13
public==2019.4.13
pyrsistent==0.15.2
Python-EasyConfig==0.1.7
pytz==2019.1
PyYAML==5.1.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
Resource==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
